requirements.txt
aiohttp==3.8.3
aiohttp-apispec==2.2.3
aiohttp-session==2.12.0
aiosignal==1.3.1
apispec==3.3.2
async-timeout==4.0.2
attrs==22.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
exceptiongroup==1.1.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
packaging==23.0
pluggy==1.0.0
pycparser==2.21
pytest==7.2.1
pytest-aiohttp==0.3.0
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.4.0
webargs==5.5.3
yarl==1.8.2